requirements.txt
tqdm
Pillow
pydub
pydantic
mortis